+++ src/constants.ts
@@ -1,0 +1,2 @@
+
+export const SHIRT_ARG_LENGTH = 2;
