+++ src/strictOptions.ts
@@ -1,0 +1,141 @@
+import { findName } from './helpers';
+import { SHIRT_ARG_LENGTH } from './constants';
+import argv from './argv';
+
+/**
+ * Strict reading of flags and options.
+ *
+ *     command-line-app --output test -p es2015,react --verbose
+ *
+ *     strictOptions([
+ *         ["another"]      - for flags array is same as object notation
+ *         "verbose"        - only one variant of name
+ *     ], [
+ *         {"output": "o"}, - fullname and shirtname
+ *         ["plugins", "p"] - fullname and shirtname for array
+ *     ])
+ *
+ * @param  {...Object} flagsNames   array of tokens
+ * @param  {...Object} optionsNames array of tokens
+ * @return {Object}    fullname-value pairs
+ */
+export function strictOptions(flagsNames, optionsNames) {
+
+	if (!argv.length) {
+		return {};
+	}
+
+	const options = {};
+
+	let argument = argv[0];
+
+	while (
+		argv.length
+		&& (
+			argument.indexOf('--') === 0
+			|| argument.indexOf('-') === 0 && argument.length === SHIRT_ARG_LENGTH
+		)
+	) {
+
+		const sourceKey = argv.shift().replace(/^(--|-)/, '');
+		const flagKey = findName(sourceKey, flagsNames);
+		const optionKey = findName(sourceKey, optionsNames);
+
+		if (!flagKey && !optionKey) {
+			throw new Error(`Unexpected key "${sourceKey}".`);
+		}
+
+		if (!flagKey && optionKey && !argv.length) {
+			throw new Error('Unexpected end of arguments.');
+		}
+
+		let value: string|string[] = argv[0];
+
+		if (optionKey && value.indexOf('--') !== 0
+			&& (value.indexOf('-') !== 0 && value.length !== SHIRT_ARG_LENGTH)
+		) {
+
+			argv.shift();
+
+			if (optionKey.isArray) {
+				value = value.split(',').map(element => element.replace(/^['"]|["']$/g, ''));
+			} else {
+				value = value.replace(/^['"]|["']$/g, '');
+			}
+
+			options[optionKey.name] = value;
+
+		} else
+		if (flagKey) {
+
+			options[flagKey.name] = true;
+
+		} else {
+			throw new Error(`Unexpected key "${value}".`);
+		}
+
+		argument = argv[0];
+	}
+
+	return options;
+}
+
+/**
+ * Strict reading of options with equal sign.
+ * If option is provided without value it will interpreted as `true`.
+ *
+ *     command-line-app --output=test -p=es2015,react --verbose
+ *
+ *     strictOptionsEqual(
+ *         {"output": "o"},  - fullname and shirtname
+ *         ["plugins", "p"], - fullname and shirtname for array
+ *         "verbose"         - only one variant of name
+ *     )
+ *
+ * @param  {...Object} names array of tokens
+ * @return {Object}    fullname-value pairs
+ */
+export function strictOptionsEqual(...names) {
+
+	if (!argv.length) {
+		return {};
+	}
+
+	const options = {};
+
+	let argument = argv[0];
+
+	while (
+		argv.length
+		&& (
+			argument.indexOf('--') === 0
+			|| argument.indexOf('-') === 0
+		)
+	) {
+
+		const keyValue = argv.shift().replace(/^(--|-)/, '').split('=');
+		const sourceKey: string = keyValue[0];
+		let value: string|string[]|boolean = keyValue[1];
+
+		const key = findName(sourceKey, names);
+
+		if (!key) {
+			throw new Error(`Unexpected key "${sourceKey}".`);
+		}
+
+		if (typeof value === 'undefined') {
+			value = true;
+		} else
+		if (key.isArray) {
+			value = value.split(',').map(element => element.replace(/^['"]|["']$/g, ''));
+		} else {
+			value = value.replace(/^['"]|["']$/g, '');
+		}
+
+		options[key.name] = value;
+
+		argument = argv[0];
+	}
+
+	return options;
+}
