+++ tslint.json
@@ -1,3 +1,3 @@
 {
-	"extends": "tslint-config-trigen"
+	"extends": "tslint-config-trigen/base"
 }
